scraper/requirements.txt
playwright
flask
gunicorn
python-dotenv
celery
gevent
redis
pillow
boto3
